add alternative storage

+++ postgresql/psb_db_prep_stand4.sh
@@ -191,17 +191,18 @@ pg_dropcluster $PG_VERSION $PG_MAIN_CLUSTER --stop
 rm -rf /etc/postgresql/$PG_VERSION/$PG_MAIN_CLUSTER
 
 # подключить диск
-# подключить диск
-lsblk | grep "${STORAGE}"
-if [ $? -eq 0 ]; then
-    lsblk | grep "${STORAGE}"
-    if [ $? -eq 0 ]; then
-        umount /var/lib/postgresql/11/
-        parted -s /dev/${STORAGE} select && parted -s /dev/${STORAGE} rm 1
-    fi
-    parted -s /dev/${STORAGE} mklabel msdos mkpart primary xfs 0% 100%
-    mkfs -t xfs -f /dev/${STORAGE}1
-    mount /dev/${STORAGE}1 /var/lib/postgresql/11/
+if [ "$2" == "SDA" ] || [ "$3" == "SDA" ]; then
+  lsblk | grep "${STORAGE}"
+  if [ $? -eq 0 ]; then
+      lsblk | grep "${STORAGE}"
+      if [ $? -eq 0 ]; then
+          umount /var/lib/postgresql/11/
+          parted -s /dev/${STORAGE} select && parted -s /dev/${STORAGE} rm 1
+      fi
+      parted -s /dev/${STORAGE} mklabel msdos mkpart primary xfs 0% 100%
+      mkfs -t xfs -f /dev/${STORAGE}1
+      mount /dev/${STORAGE}1 /var/lib/postgresql/11/
+  fi
 fi
 
 for port in $(pg_lsclusters -h | gawk '{print $3}');
